+++ lib/carousel_state.dart
@@ -23,5 +23,15 @@ class CarouselState {
   /// The widgets count that should be shown at carousel
   int itemCount;
 
-  CarouselState(this.options);
+  /// Will be called when using pageController to go to next page or 
+  /// previous page. It will clear the autoPlay timer.
+  /// Internal use only
+  Function onResetTimer;
+
+  /// Will be called when using pageController to go to next page or 
+  /// previous page. It will restart the autoPlay timer.
+  /// Internal use only
+  Function onResumeTimer;
+
+  CarouselState(this.options, this.onResetTimer, this.onResumeTimer);
 }